+++ lib/src/db/lib_prod_dbs.dart
@@ -26,3 +26,5 @@ class DbProd extends DbEnv {
   static final utility = Db('mongodb://${DbProd._connection}/utility');
   static final weather = Db('mongodb://${DbProd._connection}/weather');
 }
+
+