+++ test/db/lib_update_dbs_test.dart
@@ -60,11 +60,11 @@ Future<void> tests() async {
 
   // await updateIsoneZonalDemand([2021], download: false);
   // await updateIsoneZonalDemand(IntegerRange(2011, 2021));
-  final months = Month(2023, 6, location: IsoNewEngland.location)
-      .upTo(Month(2023, 12, location: IsoNewEngland.location));
+  final months = Month(2023, 1, location: IsoNewEngland.location)
+      .upTo(Month(2023, 1, location: IsoNewEngland.location));
   // final months = [Month.utc(2023, 1)];
-  // await updateDaEnergyOffersIsone(months: months);
-  await updateRtEnergyOffersIsone(months: months, download: true);
+  // await updateDaEnergyOffersIsone(months: months, download: false);
+  await updateRtEnergyOffersIsone(months: months, download: false);
   // await updateMorningReport(months: months, download: true);
   // await updateSevenDayCapacityForecast(months: months);